+++ BreaksPPU/PPUPlayer/FormSettings.cs
@@ -11,6 +11,7 @@ using System.Windows.Forms;
 using System.Xml.Serialization;
 using System.IO;
 using static PPUPlayer.FormSettings;
+using SharpTools;
 
 namespace PPUPlayer
 {
@@ -72,7 +73,7 @@ namespace PPUPlayer
 			settings.ColorDebug = true;
 			settings.PpuRAWMode = false;
 			settings.FreeModeVMirroring = true;
-			settings.OAMDecay = PPUPlayerInterop.OAMDecayBehavior.Keep;
+			settings.OAMDecay = BreaksCoreInterop.OAMDecayBehavior.Keep;
 			settings.PpuNoise = 0.0f;
 
 			SaveSettings(settings);
@@ -153,7 +154,7 @@ namespace PPUPlayer
 			[XmlElement]
 			[Category("Board Features")]
 			[Description("Set one of the ways to decay OAM cells.")]
-			public PPUPlayerInterop.OAMDecayBehavior OAMDecay { get; set; }
+			public BreaksCoreInterop.OAMDecayBehavior OAMDecay { get; set; }
 
 			[XmlElement]
 			[Category("Board Features")]
