+++ BreaksPPU/PPUPlayer/PPUPlayer.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PPUPlayer", "PPUPlayer.cspr
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PPUSim", "..\PPUSim\Scripts\VS2022\PPUSim.vcxproj", "{EBD9B3EB-3C04-43ED-B454-E9442B21F5A0}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PPUPlayerInterop", "..\PPUPlayerInterop\PPUPlayerInterop.vcxproj", "{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BaseBoardLib", "..\..\Common\BaseBoardLib\Scripts\VS2022\BaseBoardLib.vcxproj", "{36F535AD-B87B-4F4D-A5F9-0F2377FBA7EA}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BaseLogicLib", "..\..\Common\BaseLogicLib\Scripts\VS2022\BaseLogicLib.vcxproj", "{11AAD192-46EB-4D5D-B81F-BCEE11D6AF8E}"
@@ -93,30 +91,6 @@ Global
 		{EBD9B3EB-3C04-43ED-B454-E9442B21F5A0}.Release|x64.Build.0 = Release|x64
 		{EBD9B3EB-3C04-43ED-B454-E9442B21F5A0}.Release|x86.ActiveCfg = Release|Win32
 		{EBD9B3EB-3C04-43ED-B454-E9442B21F5A0}.Release|x86.Build.0 = Release|Win32
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug Breaks|Any CPU.ActiveCfg = Debug|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug Breaks|Any CPU.Build.0 = Debug|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug Breaks|x64.ActiveCfg = Debug|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug Breaks|x64.Build.0 = Debug|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug Breaks|x86.ActiveCfg = Debug|Win32
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug Breaks|x86.Build.0 = Debug|Win32
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug|Any CPU.Build.0 = Debug|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug|x64.ActiveCfg = Debug|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug|x64.Build.0 = Debug|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug|x86.ActiveCfg = Debug|Win32
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Debug|x86.Build.0 = Debug|Win32
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release Breaks|Any CPU.ActiveCfg = Release|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release Breaks|Any CPU.Build.0 = Release|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release Breaks|x64.ActiveCfg = Release|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release Breaks|x64.Build.0 = Release|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release Breaks|x86.ActiveCfg = Release|Win32
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release Breaks|x86.Build.0 = Release|Win32
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release|Any CPU.ActiveCfg = Release|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release|Any CPU.Build.0 = Release|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release|x64.ActiveCfg = Release|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release|x64.Build.0 = Release|x64
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release|x86.ActiveCfg = Release|Win32
-		{22245846-1FCB-484C-AD7E-AB6C7B7F1D37}.Release|x86.Build.0 = Release|Win32
 		{36F535AD-B87B-4F4D-A5F9-0F2377FBA7EA}.Debug Breaks|Any CPU.ActiveCfg = Debug|x64
 		{36F535AD-B87B-4F4D-A5F9-0F2377FBA7EA}.Debug Breaks|Any CPU.Build.0 = Debug|x64
 		{36F535AD-B87B-4F4D-A5F9-0F2377FBA7EA}.Debug Breaks|x64.ActiveCfg = Debug|x64
