--- BreaksPPU/PPUPlayerInterop/Readme.md
@@ -1,3 +1,0 @@
-# PPUPlayerInterop
-
-Interop DLL for PPUPlayer, which contains PPUSim and auxiliary logic to maintain the PPU environment.
